Split cleaning and reorder compilation and linting
There is now a separate target to allow for `make clean`ing of linter's
stuff only so that you can re-lint without recompiling.
Also, `make` now first compiles the project and then lints. It is more
important to have code that compiles than code that lints (although the
latter is still super important).