Navigation basic skeleton (#48)
* Add navigation class skeleton

* Update style

* Add main navigation thread

* Add navigation files

* Update description, name and style

* Added namespace for navigation files

* Update constructor method to support Vector class

* Update navigation files

* Update reference to be a constant reference

* Refactor some declarations from cpp file to hpp file

* Add argument for update function

* Refactor function calls

* Implement basic logic and data flow in the nav thread
The main nav thread now reads and writes the shared data. There is also
a basic skeleton for triggering appropriate nav calculations.

* Implement basic skeleton of navigation algorithm
This skeleton should be able to process accelerometer data correctly,
producing velocity and displacement. It doesn't use filtering and
ignores all other data so it it will be very imprecise. It is only
useful to test the data flow in our system.

* Nav refractor kalman (#47)

* Fill up description of Kalman class

* Refactor Kalman Class

* Revert "Refactor Kalman Class"

This reverts commit 62535db159ed951439f7bd030410fbacacef0acc.

* Remove estimation error from the constructor

* Fix issues

* Add comments and fix types
Acceleration and angular velocity are expected to be negative at times.
While velocity and displacement should not be negative during the
hyperloop run, it could theoretically happen. Anyway, it's better not to
mix signed and unsigned integers so I made them all signed.

* Reformat the license comment

* Merge changes from develop

To get up-to-date version of the main `main.cpp`

* Add navigation thread to the main `main.cpp`